poprawki countable i count

+++ backups.php
@@ -2,19 +2,19 @@
 /**
  * All Backups
  *
- * Displays all available page backups. 	
+ * Displays all available page backups.
  *
  * @package GetSimple
  * @subpackage Backups
  * @link http://get-simple.info/docs/restore-page-backup
  */
- 
+
 // Setup inclusions
 $load['plugin'] = true;
 
 // Include common.php
 include('inc/common.php');
-	
+
 // Variable settings
 login_cookie_check();
 $path = GSBACKUPSPATH.'pages/';
@@ -28,11 +28,11 @@ if (isset($_GET['deleteall'])){
 	if (!defined('GSNOCSRF') || (GSNOCSRF == FALSE) ) {
 		$nonce = $_GET['nonce'];
 		if(!check_nonce($nonce, "deleteall")) {
-			die("CSRF detected!");	
+			die("CSRF detected!");
 		}
 	}
 	$filenames = getFiles($path);
-	
+
 	foreach ($filenames as $file) {
 		if (file_exists($path . $file) ) {
 			if (isFile($file, $path, 'bak')) {
@@ -40,7 +40,7 @@ if (isset($_GET['deleteall'])){
 			}
 		}
 	}
-	
+
 	$success = i18n_r('ER_FILE_DEL_SUC');
 }
 
@@ -49,13 +49,13 @@ if (isset($_GET['deleteall'])){
 $filenames = getFiles($path);
 $count="0";
 $pagesArray_tmp = array();
-$pagesSorted=array(); 
+$pagesSorted=array();
 
-if (count($filenames) != 0) 
-{ 
-	foreach ($filenames as $file) 
+if (count($filenames) != 0)
+{
+	foreach ($filenames as $file)
 	{
-		if (isFile($file, $path, 'bak')) 
+		if (isFile($file, $path, 'bak'))
 		{
 			$data = getXML($path .$file);
 			$status = $data->menuStatus;
@@ -68,15 +68,15 @@ if (count($filenames) != 0)
 	$pagesSorted = subval_sort($pagesArray_tmp,'title');
 }
 
-if (count($pagesSorted) != 0) 
-{ 
-	foreach ($pagesSorted as $page) 
-	{					
+if (is_countable($pagesSorted) && count($pagesSorted) != 0)
+{
+	foreach ($pagesSorted as $page)
+	{
 		$counter++;
 		$table .= '<tr id="tr-'.$page['url'] .'" >';
-		
+
 		if ($page['title'] == '' ) { $page['title'] = '[No Title] &nbsp;&raquo;&nbsp; <em>'. $page['url'] .'</em>'; }
-		
+
 		$table .= '<td class="pagetitle"><a title="'.i18n_r('VIEWPAGE_TITLE').' '. var_out($page['title']) .'" href="backup-edit.php?p=view&amp;id='. $page['url'] .'">'. cl($page['title']) .'</a></td>';
 		$table .= '<td style="width:80px;text-align:right;" ><span>'. shtDate($page['date']) .'</span></td>';
 		$table .= '<td class="delete" ><a class="delconfirm" title="'.i18n_r('DELETEPAGE_TITLE').' '. var_out($page['title']) .'?" href="backup-edit.php?p=delete&amp;id='. $page['url'] .'&amp;nonce='.get_nonce("delete", "backup-edit.php").'">&times;</a></td>';
@@ -84,18 +84,18 @@ if (count($pagesSorted) != 0)
 	}
 }
 
-get_template('header', cl($SITENAME).' &raquo; '.i18n_r('BAK_MANAGEMENT')); 
+get_template('header', cl($SITENAME).' &raquo; '.i18n_r('BAK_MANAGEMENT'));
 
 ?>
-	
+
 <?php include('template/include-nav.php'); ?>
 
 <div class="bodycontent clearfix">
-	
+
 	<div id="maincontent">
 		<div class="main" >
 			<h3 class="floated"><?php i18n('PAGE_BACKUPS');?></h3>
-			
+
 			<?php if ($counter > 0) { ?>
 				<div class="edit-nav clearfix" ><a href="#" id="filtertable" accesskey="<?php echo find_accesskey(i18n_r('FILTER'));?>" ><?php i18n('FILTER'); ?></a> <a href="backups.php?deleteall&amp;nonce=<?php echo get_nonce("deleteall"); ?>" title="<?php i18n('DELETE_ALL_BAK');?>" accesskey="<?php echo find_accesskey(i18n_r('ASK_DELETE_ALL'));?>" class="confirmation"  ><?php i18n('ASK_DELETE_ALL');?></a></div>
 				<div id="filter-search">
@@ -108,11 +108,11 @@ get_template('header', cl($SITENAME).' &raquo; '.i18n_r('BAK_MANAGEMENT'));
 			<?php  } else { ?>
 				<div class="clearfix" style="height:40px;"></div>
 			<?php  }	?>
-		
+
 			<p><em><b><span id="pg_counter"><?php echo $counter; ?></span></b> <?php echo i18n_r('PAGE_BACKUPS');?></em></p>
 		</div>
 	</div>
-	
+
 	<div id="sidebar" >
 		<?php include('template/sidebar-backups.php'); ?>
 	</div>
